+++ src/components/CardAudio.js
@@ -6,7 +6,7 @@ export default function CardAudio() {
 
   const playSound = async () => {
     const { sound } = await Audio.Sound.createAsync(
-      require('../assets/audios/RELINCHO.mp3')
+      require('../assets/audios/RELINCHO.mp3'), // FIXME: A better way yo handle this constant
     )
     setSound(sound)
     await sound.playAsync()