+++ src/navigation/StartStack.js
@@ -1,26 +1,24 @@
 import { createStackNavigator } from '@react-navigation/stack'
-import HOME from '../screens/Home'
+import Home from '../screens/Home'
 import Navigation from './Navigation'
 import Tutorial from '../screens/Tutorial'
-import EditCronograma from '../screens/EditCronograma'
 import WorkView from '../screens/WorkView'
-import DetailCronograma from '../screens/DetailCronograma'
-import AddCronograma from '../screens/AddCronograma'
 
-import { SCHEDULE_DETAIL } from '../strings'
+import { HOME, NAVIGATION, TUTORIAL, WORK } from '../screens'
 
 const Stack = createStackNavigator()
 
 function StartStack() {
   return (
-    <Stack.Navigator>
-      <Stack.Screen name="HOME" component={HOME} />
-      <Stack.Screen name="CRONOS" component={Navigation} />
-      <Stack.Screen name="Tutorial" component={Tutorial} />
-      <Stack.Screen name="Editar" component={EditCronograma} />
-      <Stack.Screen name="TRABAJAR" component={WorkView} />
-      <Stack.Screen name={SCHEDULE_DETAIL} component={DetailCronograma} />
-      <Stack.Screen name="AddCronograma" component={AddCronograma} />
+    <Stack.Navigator
+      screenOptions={{
+        headerShown: false,
+      }}
+    >
+      <Stack.Screen name={HOME} component={Home} />
+      <Stack.Screen name={NAVIGATION} component={Navigation} />
+      <Stack.Screen name={TUTORIAL} component={Tutorial} />
+      <Stack.Screen name={WORK} component={WorkView} />
     </Stack.Navigator>
   )
 }