+++ src/screens/AddCard.js
@@ -1,27 +1,35 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react'
+import { Button, Image, StyleSheet, Text, TextInput } from 'react-native'
+import { useNavigation } from '@react-navigation/native'
+import * as ImagePicker from 'expo-image-picker'
+import * as MediaLibrary from 'expo-media-library'
+import { LinearGradient } from 'expo-linear-gradient'
+
+import { saveCard, openDatabase } from '../model'
 import {
-  Button,
-  Image,
-  StyleSheet,
-  Text,
-  TextInput,
-} from 'react-native';
-import * as ImagePicker from 'expo-image-picker';
-import * as MediaLibrary from 'expo-media-library';
-import PropTypes from 'prop-types';
-import { LinearGradient } from 'expo-linear-gradient'; 
+  ENTER_DESCRIPTION,
+  ENTER_TITLE,
+  SAVE,
+  SELECT_IMAGE,
+  CARD_SAVED,
+} from '../strings'
+import { CARDS_LIST } from '../screens'
+import {
+  BACKGROUND_GRADIENT_1,
+  BACKGROUND_GRADIENT_2,
+  PRIMARY_BUTTON_COLOR,
+} from '../styles'
 
-import { saveCard, openDatabase } from '../model';
-
-function AddCard({ navigation }) {
-  const [image, setImage] = useState(null);
-  const [title, setTitle] = useState(null);
-  const [description, setDescription] = useState(null);
-  const [permissionResponse, requestPermission] = MediaLibrary.usePermissions();
+function AddCard() {
+  const navigation = useNavigation()
+  const [image, setImage] = useState(null)
+  const [title, setTitle] = useState(null)
+  const [description, setDescription] = useState(null)
+  const [permissionResponse, requestPermission] = MediaLibrary.usePermissions()
 
   useEffect(() => {
-    requestPermission();
-  }, [requestPermission]);
+    requestPermission()
+  }, [requestPermission])
 
   const onPressImage = async () => {
     let result = await ImagePicker.launchImageLibraryAsync({
@@ -29,43 +37,48 @@ function AddCard({ navigation }) {
       allowsEditing: true,
       aspect: [4, 3],
       quality: 1,
-    });
+    })
     if (!result.canceled) {
-      setImage(result.assets[0].uri);
+      setImage(result.assets[0].uri)
     }
-  };
+  }
 
   const onPressSave = async () => {
     if (!permissionResponse?.granted) {
-      console.log('Sorry, we need media permissions');
-      return;
+      console.log('Sorry, we need media permissions')
+      return
     }
 
     try {
       // Request device storage access permission
-      const { status } = await MediaLibrary.requestPermissionsAsync();
+      const { status } = await MediaLibrary.requestPermissionsAsync()
       if (status === 'granted') {
-        const asset = await MediaLibrary.createAssetAsync(image);
-        const db = openDatabase();
+        const asset = await MediaLibrary.createAssetAsync(image)
+        const db = openDatabase()
         const card = {
           title: title,
           description: description,
           audio: '',
           image: asset.uri,
-        };
-        saveCard(db, card);
+        }
+        saveCard(db, card)
         // FIXME: Show something to the user
-        navigation.navigate('TARJETAS');
-        console.log('Image successfully saved');
+        navigation.navigate(CARDS_LIST, {
+          action: {
+            success: true,
+            message: CARD_SAVED,
+          },
+        })
+        console.log('Image successfully saved')
       }
     } catch (error) {
-      console.log(error);
+      console.log(error)
     }
-  };
+  }
 
   return (
-    <LinearGradient 
-      colors={['rgb(207,226,136)', 'rgb(45,105,129)']}
+    <LinearGradient
+      colors={[BACKGROUND_GRADIENT_1, BACKGROUND_GRADIENT_2]} // FIXME: Replace for
       style={styles.container}
     >
       <Text>Título</Text>
@@ -73,31 +86,31 @@ function AddCard({ navigation }) {
         style={styles.input}
         onChangeText={setTitle}
         value={title}
-        placeholder="Ingrese el título de la tarjeta"
+        placeholder={ENTER_TITLE}
       />
       <Text>Descripción</Text>
       <TextInput
         style={styles.input}
         onChangeText={setDescription}
         value={description}
-        placeholder="Ingrese una descripción"
+        placeholder={ENTER_DESCRIPTION}
       />
       <Button
-        title="Seleccionar imagen"
-        color="#2D6981"
+        title={SELECT_IMAGE}
+        color={PRIMARY_BUTTON_COLOR}
         onPress={onPressImage}
       />
       {image && (
         <Image source={{ uri: image }} style={{ width: 200, height: 200 }} />
       )}
-      <Button title="Guardar" color="#2D6981" onPress={() => onPressSave()} />
+      <Button
+        title={SAVE}
+        color={PRIMARY_BUTTON_COLOR}
+        onPress={() => onPressSave()}
+      />
     </LinearGradient>
-  );
+  )
 }
-
-AddCard.propTypes = {
-  navigation: PropTypes.object.isRequired,
-};
 
 const styles = StyleSheet.create({
   container: {
@@ -110,6 +123,6 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     padding: 10,
   },
-});
+})
 
-export default AddCard;
+export default AddCard